Improve directory structure.
