+++ graylog2-web-interface/src/components/streams/StreamsOverview/ColumnRenderers.tsx
@@ -23,32 +23,19 @@ import { Label } from 'components/bootstrap';
 import { Link } from 'components/common/router';
 import Routes from 'routing/Routes';
 import type { ColumnRenderers } from 'components/common/EntityDataTable';
-import IndexSetCell from 'components/streams/StreamsOverview/IndexSetCell';
-import ThroughputCell from 'components/streams/StreamsOverview/ThroughputCell';
+import IndexSetCell from 'components/streams/StreamsOverview/cells/IndexSetCell';
+import ThroughputCell from 'components/streams/StreamsOverview/cells/ThroughputCell';
 import type { IndexSet } from 'stores/indices/IndexSetsStore';
 import { CountBadge } from 'components/common';
 import useExpandedSections from 'components/common/EntityDataTable/hooks/useExpandedSections';
 
-import StatusCell from './StatusCell';
+import StatusCell from './cells/StatusCell';
 
 const DefaultLabel = styled(Label)`
   display: inline-flex;
   margin-left: 5px;
   vertical-align: inherit;
 `;
-
-const StreamRuleCounter = ({ stream }: { stream: Stream }) => {
-  const buttonRef = useRef();
-  const { toggleSection } = useExpandedSections();
-
-  const toggleRulesSection = useCallback(() => toggleSection(stream.id, 'rules'), [stream.id, toggleSection]);
-
-  return (
-    <CountBadge onClick={toggleRulesSection} ref={buttonRef}>
-      {stream.rules.length}
-    </CountBadge>
-  );
-};
 
 const customColumnRenderers = (indexSets: Array<IndexSet>): ColumnRenderers<Stream> => ({
   title: {
@@ -72,7 +59,7 @@ const customColumnRenderers = (indexSets: Array<IndexSet>): ColumnRenderers<Stre
     staticWidth: 100,
   },
   rules: {
-    renderCell: (stream) => <StreamRuleCounter stream={stream} />,
+    renderCell: (stream) => <StreamRuleCell stream={stream} />,
     staticWidth: 50,
   },
 });
