+++ graylog2-web-interface/src/components/streams/StreamsOverview/cells/IndexSetCell.tsx
