+++ graylog2-web-interface/src/components/streams/StreamsOverview/cells/ThroughputCell.tsx
